# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code Story API",
...
    "contact": {
      "name": "Code Story Support",
      "url": "https://codestory.dev/support",
      "email": "api-support@codestory.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/me/api-keys": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Api Keys",
        "description": "List user's API keys.\n\nArgs:\n    user: Authenticated user\n    db: Database session\n\nReturns:\n    List of API keys (without the full key)",
        "operationId": "list_api_keys_api_users_me_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Api Users Me Api Keys Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key.\n\nThe full key is only returned once on creation.\nStore it securely - it cannot be retrieved again.\n\nArgs:\n    request: API key creation parameters\n    user: Authenticated user\n    db: Database session\n\nReturns:\n    Created API key with full key visible",
        "operationId": "create_api_key_api_users_me_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/me/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key.\n\nArgs:\n    key_id: API key ID to delete\n    user: Authenticated user\n    db: Database session\n\nRaises:\n    NotFoundError: If key doesn't exist or user doesn't own it",
        "operationId": "delete_api_key_api_users_me_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/api-keys": {
      "post": {
        "tags": [
          "api-keys",
          "API Keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for the authenticated user.\n\nThe returned key is only shown once and cannot be retrieved later.\nStore it securely.",
        "operationId": "create_api_key_api_api_keys_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "api-keys",
          "API Keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys for the authenticated user.",
        "operationId": "list_api_keys_api_api_keys_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return active keys",
              "default": true,
              "title": "Active Only"
            },
            "description": "Only return active keys"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/{key_id}": {
      "get": {
        "tags": [
          "api-keys",
          "API Keys"
        ],
        "summary": "Get Api Key",
        "description": "Get details of a specific API key.",
        "operationId": "get_api_key_api_api_keys__key_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/codestory__api__routers__api_keys__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "api-keys",
          "API Keys"
        ],
        "summary": "Update Api Key",
        "description": "Update an API key (name, active status, or rate limit).",
        "operationId": "update_api_key_api_api_keys__key_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "rate_limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rate Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/codestory__api__routers__api_keys__APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "api-keys",
          "API Keys"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key (permanently).",
        "operationId": "delete_api_key_api_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Name for the API key"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "Permissions for this key (read, write, admin)",
            "default": [
              "read"
            ]
          },
          "expires_in_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days",
            "description": "Days until expiration (null = never)"
          },
          "rate_limit": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 10.0,
            "title": "Rate Limit",
            "description": "Requests per hour limit",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreate",
        "description": "Request body for creating an API key."
      },
      "APIKeyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreateRequest",
        "description": "Request to create a new API key."
      },
      "APIKeyCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "rate_limit": {
            "type": "integer",
            "title": "Rate Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The API key (only shown once)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "permissions",
          "rate_limit",
          "is_active",
          "created_at",
          "expires_at",
          "last_used_at",
          "key"
        ],
        "title": "APIKeyCreated",
        "description": "Response when creating a new API key (includes the actual key).",
        "example": {
          "id": 1,
          "name": "CI/CD Pipeline Key",
          "key": "cs_a1b2c3d4e5f6...",
          "permissions": [
            "read",
            "write"
          ],
          "rate_limit": 1000,
          "is_active": true,
          "created_at": "2025-01-15T10:30:00Z",
          "expires_at": null,
          "last_used_at": null
        }
      },
      "APIKeyCreatedResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key",
          "prefix",
          "created_at"
        ],
        "title": "APIKeyCreatedResponse",
        "description": "Response with full API key (only shown once)."
      },
      "APIKeyList": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/codestory__api__routers__api_keys__APIKeyResponse"
            },
            "type": "array",
            "title": "Keys"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "keys",
          "total"
        ],
        "title": "APIKeyList",
        "description": "Response body for listing API keys."
      },
      "APIKeyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "created_at",
          "last_used_at",
          "is_active"
        ],
        "title": "APIKeyResponse",
        "description": "API key response (key shown only on creation).",
        "example": {
          "id": 1,
          "name": "CI/CD Pipeline Key",
          "permissions": [
            "read",
            "write"
          ],
          "rate_limit": 1000,
          "is_active": true,
          "created_at": "2025-01-15T10:30:00Z",
          "expires_at": null,
          "last_used_at": "2025-01-15T12:45:00Z"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "codestory__api__routers__api_keys__APIKeyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "rate_limit": {
            "type": "integer",
            "title": "Rate Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "permissions",
          "rate_limit",
          "is_active",
          "created_at",
          "expires_at",
          "last_used_at"
        ],
        "title": "APIKeyResponse",
        "description": "Response body for API key operations."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/auth/login or Supabase Auth"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key from /api/api-keys (format: cs_<hex>)"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.codestory.dev",
      "description": "Production server"
    },
    {
      "url": "https://api-staging.codestory.dev",
      "description": "Staging server"
    }
  ]
}